updating workflow to better integration test
Bottom up integration testing runs lowest integrated units first and builds to top functionality.  This change reverses the order of the test suite, running the tests with the fewest dependencies first and the most last.  This should make debugging easier because with the old system, a workflow will stop if the first test (big bang) fails, not showing the user what part of the code fails.